requirements.txt
astroid==2.11.7
dill==0.3.5.1
docopt==0.6.2
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
paho-mqtt==1.6.1
platformdirs==2.5.2
pylint==2.14.5
tomli==2.0.1
tomlkit==0.11.1
typed-ast==1.5.4
typing_extensions==4.3.0
wrapt==1.14.1
